requirements.txt
Django==2.0.1
django-model-utils==3.1.1
imgkit==0.1.8
